Add a guard against creating an octoprint backup while the printer is running.
octopi has a bunch of warnings about avoiding doing things on the
raspberry pi while a print is running to prevent the print from being
ruined. I don't know how true it is, but it seems reasonable that
creating a backup could overload the raspberry pi and screw up the
print.